feat: Bad Pixel Mask
This introduces a new attribute called 'mask' to the 'Image' object.

One can write the mask oneself to the image, but it is also automatically raised by the CleanBadPixels block with the bad pixel map.

The mask is then given to the 'aperture_photometry' function of photutils, which up to now was left as 'None'.

In summary, this feature introduces an alternative to the interpolation method of cleaning bad pixels used before.